Modified the structure of ActivePiece class
This new structure was designed to allow for the getDroppedObj()
function, which will be used in the future to render a shadow
of the active piece to show where it would land if the user drops
it all the way down. The changes in ActivePiece include splitting
drop() into two different functions, changing where lastDropTime
gets updated, and adding comments.