Revamped piece rotation system
Changed around a lot of the ActivePiece's internal rotation code
to make all the pieces rotate as they do in the real Tetris. This
involved making a "pivot" point for the ActivePiece, and the piece
is rotated around this point. Also, a rotation "limit" was added
and depends on the piece type, which lets specific pieces (such
as the long cyan piece and the two s-like pieces) only move back
and forth between two rotations.

This code is a little ugly for now, but I'll fix it soon.