Server now sends table data to all clients
The server sends a 'packet' object 30 times a second. This object
consists of arrays of the keys and values of the gamedata map,
so each client knows who is connected and what their game tables
look like. This probably isn't the most efficient way of
calculating and sending the game data, since the server has to
recalculate the players and tables arrays for each update, so
this method might be replaced in the future.