Restructured index.html
This new index page has separate canvasses for all the different
parts of the game UI, which makes it much easier to support a
variety of screen sizes going forward.

The opponentCanvas needs some work though, since its actual width
doesn't match its clientWidth, and also because it gets absurdly
large when the page is fullscreen.

This page is NOT functional yet, and is just a proof of concept.