Fixed bug with opponent screens not always displaying
The opponent canvas on some clients was not properly resizing after
multiple users joined and left the webpage -- this was caused by
getNumOpponents() returning a smaller value than the actual number
of current opponents. The root problem was that the server would
announce when a player left, even if that player didn't actually
join the game by entering a username and playing; the client would
then incorrectly decrease its opponent count, leading to the wrong
getNumOpponents() result.