Fixed minor bug when switching hosts
If the current host closes the webpage before starting the game,
then the server automatically selects another host, but
the new host previously wouldn't have the option to start the game.
This bug fix makes the host screen appear for a new host, so the
user can start the game after a previous host disconnects.