Added game over screen on client
The new screen shows up once the server signals that the game
ended, and displays who the winner is. If the current user is
the game host, then they will see a button to play again
(which isn't implemented yet).

The system for using a newline character with Node.textContent
comes from
https://stackoverflow.com/questions/9980416/how-can-i-insert-new-line-carriage-returns-into-an-element-textcontent
.